fix tags editing improve error handling

+++ portal/endpoints/tags.py
@@ -1,8 +1,10 @@
 import json
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
-from django.http import JsonResponse, HttpResponse
+from django.http import HttpResponseBadRequest, JsonResponse, HttpResponse
 from django.views.decorators.http import require_POST, require_GET
+from django.db import transaction
+from portal.endpoints.common import json_load_body
 
 from portal.models import *
 
@@ -25,17 +27,22 @@ def case_tags(request, case):
     and tags specific to the current case.
     '''
     case = get_object_or_404(Case, id=case)
-    return JsonResponse({"tags": [tag.name for tag in Tag.objects.all()], "case_tags": [tag.name for tag in case.tags.all()]}, safe=False)
+    return JsonResponse({"tags": [tag.name for tag in Tag.objects.all()], 
+                         "case_tags": [tag.name for tag in case.tags.all()]}, safe=False)
 
 
 @login_required
 @require_POST
+@transaction.atomic
 def update_case_tags(request, case):
     '''
     Updates tags for a given case and returns response.ok if successful
     '''
-    body = json.loads(request.body.decode('utf-8'))
     case = get_object_or_404(Case, id=case)
+    body = json_load_body(request)
+    if 'tags' not in body:
+        return HttpResponseBadRequest()
+
     tags = body['tags']
 
     # clear old tags from the case, but keep the tags in db
@@ -44,28 +51,29 @@ def update_case_tags(request, case):
         case.tags.remove(old_tag)
 
     for tag in tags:
-        existing_tag = Tag.objects.filter(name=tag).first()
-        if(existing_tag is None):
-            new_tag = Tag(name=tag)
-            new_tag.save()
-            case.tags.add(new_tag)
-            # new_tag.cases.add(case)                
-        else:
-            case.tags.add(existing_tag)
-            # existing_tag.cases.add(case)               
+        the_tag = Tag.objects.filter(name=tag).first()
+        if the_tag is None:
+            the_tag = Tag(name=tag)
+            the_tag.save()
+        case.tags.add(the_tag)
+        
         
     return HttpResponse() 
 
  
 @login_required
 @require_POST
+@transaction.atomic
 def update_tags(request):
     '''
     Delete selected tags and returns response.ok if successful
     '''
-    body = json.loads(request.body.decode('utf-8'))
-    tags = body['tags']
-    tags = [ get_object_or_404(Tag, name=tag_name) for tag_name in tags ]
+    body = json_load_body(request)
+
+    if 'tags' not in body:
+        return HttpResponseBadRequest()
+    
+    tags = [ get_object_or_404(Tag, name=tag_name) for tag_name in body['tags'] ]
     for tag in tags:
         tag.delete()
